Move AccountActions component into its own file
